back_end/python-bot/requirements.txt
bs4
praw
google-cloud-storage
numpy
selenium
googletrans
mechanize